+++ app/javascript/lca/utils/calculated/weapons/overwhelming.js
@@ -5,12 +5,15 @@ export function weaponOverwhelming(character: Character, weapon: fullWeapon) {
   let bonus = weapon.bonus_overwhelming
 
   if (weapon.tags.includes('subtle')) return 0 + bonus
-  if (weapon.tags.includes('elemental bolt'))
+  if (
+    weapon.tags.includes('elemental bolt') ||
+    weapon.tags.includes('crypt bolt')
+  )
     return 1 + character.essence + bonus
 
   if (weapon.tags.includes('balanced')) bonus += 1
 
-  const overwhelmingTag = weapon.tags.find(t => t.startsWith('overwhelming+'))
+  const overwhelmingTag = weapon.tags.find((t) => t.startsWith('overwhelming+'))
   if (overwhelmingTag) {
     let theBonus = overwhelmingTag.substr(13)
     bonus += theBonus.includes('essence')